@@ -5,34 +5,36 @@
   "dependencies": {
     "@fortawesome/fontawesome-free-brands": "^5.0.13",
     "@fortawesome/fontawesome-free-solid": "^5.0.13",
-    "@fortawesome/fontawesome-svg-core": "^1.2.8",
-    "@fortawesome/react-fontawesome": "^0.1.3",
-    "classnames": "^2.2.6",
+    "@fortawesome/fontawesome-svg-core": "^6.5.1",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "classnames": "^2.5.1",
     "cross-env": "^7.0.3",
-    "date-fns": "2.0.0-alpha.25",
-    "gatsby": "^2.0.19",
-    "gatsby-image": "^2.0.15",
-    "gatsby-plugin-feed": "^2.2.0",
-    "gatsby-plugin-google-analytics": "^2.0.19",
-    "gatsby-plugin-manifest": "^2.0.5",
-    "gatsby-plugin-offline": "^2.0.11",
-    "gatsby-plugin-react-helmet": "^3.0.5",
-    "gatsby-plugin-react-leaflet": "^2.0.8",
-    "gatsby-plugin-sharp": "^2.0.7",
-    "gatsby-source-filesystem": "^2.0.12",
-    "gatsby-transformer-json": "^2.1.11",
-    "gatsby-transformer-remark": "^2.1.19",
-    "gatsby-transformer-sharp": "^2.1.4",
+    "date-fns": "^3.1.0",
+    "enzyme-adapter-react-16": "^1.15.7",
+    "eslint-plugin-standard": "^5.0.0",
+    "gatsby": "^5.3.3",
+    "gatsby-image": "^3.11.0",
+    "gatsby-plugin-feed": "^5.13.0",
+    "gatsby-plugin-manifest": "^5.13.0",
+    "gatsby-plugin-offline": "^6.13.0",
+    "gatsby-plugin-react-helmet": "^6.13.0",
+    "gatsby-plugin-react-leaflet": "^4.0.3",
+    "gatsby-plugin-sharp": "^5.13.0",
+    "gatsby-source-filesystem": "^5.13.0",
+    "gatsby-transformer-json": "^5.13.0",
+    "gatsby-transformer-remark": "^6.13.0",
+    "gatsby-transformer-sharp": "^5.13.0",
     "he": "^1.2.0",
     "is-url": "^1.2.4",
-    "leaflet": "^1.5.1",
-    "node-html-to-image": "^3.2.4",
-    "query-string": "^7.1.1",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-helmet": "^5.2.0",
-    "react-leaflet": "^2.3.0",
-    "react-responsive": "^6.0.1",
+    "leaflet": "^1.9.4",
+    "node-html-to-image": "^4.0.0",
+    "prop-types": "^15.8.1",
+    "query-string": "^8.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-helmet": "^6.1.0",
+    "react-leaflet": "^4.2.1",
+    "react-responsive": "^9.0.2",
     "smart-truncate": "^1.0.1"
   },
   "keywords": [
@@ -62,30 +64,28 @@
     "prebuild": "yarn makeApi",
     "build": "NODE_ENV=production gatsby build",
     "prestart": "yarn makeApi",
-    "start": "cross-env NODE_ENV=development GATSBY_GRAPHQL_IDE=playground gatsby develop",
+    "start": "gatsby develop",
     "serve": "gatsby serve"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.6",
+    "@babel/core": "^7.23.7",
+    "@babel/eslint-parser": "^7.5.4",
+    "@babel/preset-react": "^7.23.3",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^23.6.0",
-    "babel-preset-gatsby": "^0.1.4",
-    "coveralls": "^3.0.2",
-    "dotenv": "^6.1.0",
-    "enzyme": "^3.7.0",
-    "enzyme-adapter-react-16": "^1.7.0",
-    "eslint": "^5.9.0",
-    "eslint-config-standard": "^12.0.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jest": "^22.0.0",
-    "eslint-plugin-node": "^8.0.0",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-react": "^7.11.1",
-    "eslint-plugin-standard": "^4.0.0",
+    "babel-jest": "^29.7.0",
+    "babel-preset-gatsby": "^3.13.0",
+    "coveralls": "^3.1.1",
+    "dotenv": "^16.3.1",
+    "enzyme": "^3.11.0",
+    "eslint": "^8.56.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jest": "^27.6.1",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.33.2",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^23.6.0",
-    "jest-axe": "^3.2.0"
+    "jest": "^29.7.0",
+    "jest-axe": "^8.0.0"
   },
   "repository": {
     "type": "git",
