@@ -1,5 +1,5 @@
 import React from 'react'
-import { StaticQuery, graphql } from 'gatsby'
+import {graphql, useStaticQuery} from 'gatsby'
 import { compareDesc } from 'date-fns'
 import { parse } from 'date-fns/fp'
 
@@ -8,10 +8,7 @@ import { OldMeetups } from './OldMeetups'
 const parseDate = parse(new Date(), 'dd/MM/yyyy')
 
 export function OldMeetupsContainer () {
-  return (
-    <StaticQuery
-      query={graphql`
-            {
+  const data = useStaticQuery(graphql`query {
               allMarkdownRemark(
                 filter: { frontmatter: { status: { eq: "done" } } }
               ) {
@@ -42,34 +39,31 @@ export function OldMeetupsContainer () {
                       }
                     }
               } } }
-            }
-            `}
-      render={({ allMarkdownRemark: { edges } }) => {
-        const currentYear = new Date().getFullYear()
+            }`)
 
-        const allMeetups = edges
-          .map(({ node }) => {
-            const parsedDate = parseDate(node.frontmatter.date)
+  const currentYear = new Date().getFullYear()
+  const { allMarkdownRemark: { edges } } = data
+  const allMeetups = edges
+    .map(({ node }) => {
+      const parsedDate = parseDate(node.frontmatter.date)
 
-            return {
-              ...node.frontmatter,
-              parsedDate,
-              year: parsedDate.getFullYear(),
-              filename: node.parent.name,
-            }
-          })
+      return {
+        ...node.frontmatter,
+        parsedDate,
+        year: parsedDate.getFullYear(),
+        filename: node.parent.name,
+      }
+    })
+  const meetups = allMeetups
+    .filter(meetup => meetup.year === currentYear)
+    .sort((a, b) => {
+      return compareDesc(a.parsedDate, b.parsedDate)
+    })
 
-        const meetups = allMeetups
-          .filter(meetup => meetup.year === currentYear)
-          .sort((a, b) => {
-            return compareDesc(a.parsedDate, b.parsedDate)
-          })
+  const pastYearsMeetups = allMeetups.filter(meetup => meetup.year != currentYear)
+  const years = [...new Set(pastYearsMeetups.map(meetup => meetup.year))]
 
-        const pastYearsMeetups = allMeetups.filter(meetup => meetup.year != currentYear)
-        const years = [...new Set(pastYearsMeetups.map(meetup => meetup.year))]
-
-        return <OldMeetups meetups={meetups} years={years} />
-      }}
-    />
+  return (
+    <OldMeetups meetups={meetups} years={years} />
   )
 }
